Remove default-zero for `physfsrwops_seek` variable pos
This default-zero allowed the bug introduced in
68b47307b4f24c3905c740f79f759e97d5b6ff3c (and fixed in
4c371734b53bf1784d226a0f09c94fcbdbb00969) not to generate a compiler
warning for `-Wuninitialized`, since it was initialized to `0`.
However, `0` is not a correct starting value for this variable.  Remove
the initialization and require every code path to assign a meaningful
value.

Switch to use `std::in_range` to validate that no truncation occurs, and
trust that the compiler will optimize this out when the type passed to
`std::in_range` can represent all values that the argument can
represent.