Discourage spurious -Wformat warning in mission.cpp on i686-pc-linux-gnu
On i686-pc-linux-gnu, but not on x86_64-pc-linux-gnu, gcc fails to prove
the correct upper bound on the iteration of the
build_rdl_regular_level_names and similar functions, then warns that the
index could be up to (UINT_MAX / sizeof(array element)).  If such an
index occurred, the snprintf result would be truncated.  Such an index
cannot occur, because all callers set appropriate small lengths.

Since an excessively large index is impossible, discourage the warning
by advising gcc that the loop index is never above the loop iteration
count.

Reported-by: th-otto <https://github.com/dxx-rebirth/dxx-rebirth/issues/732>