Suppress spurious clang -Wbraced-scalar-init
As reported in <https://github.com/llvm/llvm-project/issues/54702>,
clang warns about use of braces in contexts where they are useful to
explicitly disallow narrowing conversions.  Such braces should not be
removed, as it would allow a future change to accidentally introduce a
narrowing conversion and not receive a diagnostic.  As of this commit,
the LLVM issue has seen no activity and there is no indication the
spurious warning will be fixed.

Add an SConf test to detect if the compiler warns for these cases, and
suppress it if so.