Remove useless end-of-buffer test
PHYSFSX_fgets never returns a pointer to the end of the buffer.  It
returns either a failure status or a pointer to the terminating null
byte in the buffer.  Since that byte must be in the buffer, it cannot be
at `line.end()`, so this test can never succeed.  Remove it.