Simplify iff_parse_bitmap
Using `bmheader.raw_data.reset` was wrong, because then both bmheader
and bm would own the data, but both would claim to own it exclusively.
Fortunately, no caller passed a non-nullptr data, so this did not cause
problems in practice.  Remove the `reset` to avoid future issues.
Simplify setup of bmheader.w, bmheader.h, since with the reset removed,
the `else` path is now guaranteed.

Remove from `iff_read_bitmap` a dangerous `bm_data = nullptr`, which
discarded the pointer without attempting to free it.  Some callers might
own the data, and need it to be freed.  Others might be borrowing the
data and need it not to be freed.  Fortunately, as with
`iff_parse_bitmap`, no caller supplied a non-nullptr value here, so no
memory was leaked by this statement.