Move segment_object_range type check into `requires()`
segment_object_range_t::construct requires a segment factory that
produces a value that can be compared to the address of a
`unique_segment &`.  This is required for the assert to compile, but
optional when assertions are disabled.  Previously, a sizeof()
expression with a similar comparison was used to enforce that
no-assertions builds still met the type requirement.  Replace this with
a `requires` check, which is clearer and produces better error messages
when it is not satisfied.