Partially revert "Read allow_nullptr from ptr/idx, not policy"
clang-16 still does not implement
<https://www.open-std.org/jtc1/sc22/wg21/docs/papers/2021/p2280r2.html>,
so it chokes on:

```
	template <typename rpolicy>
		idx(const idx<rpolicy> &rhs)
		requires(
			allow_nullptr || !rhs.allow_nullptr
		)
```

because it sees `rhs` is a reference to an unknown variable, but fails
to notice that `allow_nullptr` is a `static constexpr` member, so the
validity of the reference cannot affect the validity of `allow_nullptr`.

Rewrite this as:

```
	template <typename rpolicy>
		requires(
			allow_nullptr || !idx<rpolicy>::allow_nullptr
		)
		idx(const idx<rpolicy> &rhs) :
```

This is less maintainable since it repeats the type `idx<rpolicy>`, but
it does allow clang to see that `allow_nullptr` is not dependent on the
specific argument passed.

Reported-by: pkubaj <https://github.com/dxx-rebirth/dxx-rebirth/issues/745#issuecomment-1872624496>