Short out player_has_primary_weapon for missing weapons
For every caller, if the flag `::weapon` is not set, then no other flags
are examined.  Change player_has_primary_weapon to skip computing the
other flags if it determines that `::weapon` will not be set.

Fix up the comment explaining why the `::ammo` flag is forced for
non-vulcan weapons.