Forward unique_ptr through parse_blob_as_text_strings
Returning a local unique_ptr by-value works in gcc, but fails in clang:

```
similar/main/text.cpp:307:9: error: call to deleted constructor of 'std::unique_ptr<char[]>'
    return p;
           ^
/usr/lib/gcc/x86_64-pc-linux-gnu/13/include/g++-v13/bits/unique_ptr.h:796:7: note: 'unique_ptr' has been explicitly marked deleted here
      unique_ptr(const unique_ptr&) = delete;
      ^
```

gcc automatically treats this as a move from a local variable, and
succeeds.  clang treats this as an attempt to copy the variable, and
fails.

Rework the logic to explicitly move the variable into the helper
function, and let that helper return the variable.