Replace automatic creation of (pre)release
The original submission fails at runtime due to insufficient
permissions.  Considering its use of a continuous release from an
unrelated Github repository, which could change at any time, granting
the step permission to write to the Rebirth repository seems unwise.

Replace this step with a community-written Github Action.  Perhaps
someday Github Actions will have an Action that can create a release in
a manner that can be readily verified, at which point this can be
converted.  One such action once existed at
<https://github.com/actions/upload-release-asset>, but was archived,
abandoned, and marked unmaintained
<https://github.com/actions/upload-release-asset/issues/78>.  Despite
the claim in that issue that "a large number of really good community
Actions have appeared on the Marketplace for doing this." and
considerable time spent searching, there do not appear to be any
Marketplace-endorsed actions that can fulfill this basic need.

The Action "GH Release"
<https://github.com/marketplace/actions/gh-release> appears popular, so
even though that is marked as uncertified, provided by an untrusted
third party, and "is governed by separate terms of service, privacy
policy, and support documentation" (though neither terms of service nor
privacy policy are linked, nor anywhere to be found), switch to this
Action, since at least it can be pinned to a specific release, which
hopefully will avoid it spontaneously changing behavior in the future.