Add test for Class Template Argument Deduction
CTAD is already used in Rebirth code.  Add a test to reject compilers
which do not support it.  Include in the test a construct known to cause
gcc-11 to reject the code, since that construct already exists in the
game code, and thus gcc-11 cannot compile the game.  This test forces
a build using gcc-11 to fail early, at the configure stage, rather than
only failing when it reaches the header that uses a user defined
deduction guide for an inner class.