Fix `PHYSFSX_fgets` when read boundary ends on "\r"
`PHYSFSX_fgets` treats any of "\r", "\n", or "\r\n" as a line
terminator.  When the length of the line is such that `PHYSFSX_fgets`
places a '\r' in the last element, the prior logic would advance `p` to
point to `e`.  The subsequent call to `PHYSFS_seek` would then advance
by `p - e + 1` => `e - e + 1` => `1`.  However, the byte skipped over by
that `1` may or may not be a newline.  If it is not, then the first byte
of the following line is skipped and never seen by the caller.  Change
the test so that a "\r" in the last element leaves `p` pointing at the
"\r", so that `PHYSFS_seek` advances by `(e - 1) + e + 1` => `0`, which
leaves the unread byte to be returned by the next call to
`PHYSFSX_fgets`.