Fix physics roll
Cleanup commit 5ac56d8eb161 introduced a local cache of a variable that
may be updated by a call to `set_object_turnroll`.  The local cache was
not updated, causing use of the cached variable later in the function to
produce an incorrect result.  Fix this by rereading the variable from
the authoritative source after the call to `set_object_turnroll`.
Reduce the scope of the initial cache to reflect the limited scope in
which use of the cached variable is valid.

Fixes: 5ac56d8eb161194aa46058d3bf0ab32e4267fb35 ("Reduce scope of vms_angvec local variables")
Reported-by: ziplantil <https://github.com/dxx-rebirth/dxx-rebirth/issues/766>
Bisected-by: ziplantil <https://github.com/dxx-rebirth/dxx-rebirth/issues/766#issuecomment-2023532987>
Tested-by: ziplantil <https://github.com/dxx-rebirth/dxx-rebirth/issues/766#issuecomment-2024563975>