Return `T&` from valptridx::ptr factory iteration
Commit 03e12fedd42ac89cd62f5168badc26aad5acbcd8 ("Test for overzealous -Wdangling-reference")
added an SConf test to detect when the compiler warns for certain safe
uses of references, such as when iterating over a valptridx factory.
gcc-14 improved its warning heuristics enough to make the SConf test not
recognize -Wdangling-reference as overzealous, but not enough to
actually avoid warning about valptridx iteration.  Rework the iteration
logic to avoid the construct that triggers the gcc warning.