Factor out multiply_with_clamp_to_fixang for broader use
Various locations in physics.cpp multiply terms and produce a fixang.  If
the fixang were truncated, strange results would occur, so clamp the value
to the minimum/maximum value of a fixang before conversion.  This should
cause any out-of-range values to saturate rather than truncate and produce
a strange value.