Fix bogus objects_in gcc-14 -Wdangling-reference warning
The code was correct before, but gcc-14 warned because it detected an
anonymous temporary vmobjptridx_t was created, a method on that
temporary returned a reference, the returned reference was saved to a
longer lived variable, and then the temporary was destroyed.  If the
reference had pointed into the temporary, that would have been a
dangling reference bug.  However, the temporary only wraps a pointer to
external storage, so the reference points to that external storage, and
is not dangling.  Therefore, the warning is wrong.

Prevent the warning by restructuring `objects_in` such that the caller
can request the iterator directly return a `object &`, without the need
for an anonymous temporary vmobjptridx_t.  Since no temporary is
created, no warning is issued.

Add some `static_assert` to `segiter.h` to document the behavior of
`objects_in`.