Remove test for overzealous -Wdangling-reference
One site that only provoked gcc-13 has been reworked not to trigger
this.  Changes in gcc-14 caused this test not to recognize the compiler
as overzealous, but the underlying warning was still overzealous.  All
sites that gcc-14 warned about have been reworked to avoid the warning,
so remove the attempt to suppress it.