Rewrite score read/write to remove the dependency on structure alignment
Score read/write code depended on the C structure layout, which is
influenced by the compiler's alignment decisions.  Use of
__attribute__((packed)) or equivalent encouraged the compiler to produce
the desired layout, but did not fully guarantee the desired layout.
Some unusual platforms still laid out the structure in a way different
from DOS.  Rewrite the code to eliminate the dependency on the structure
layout.

Reported-by: mikrosk <https://github.com/dxx-rebirth/dxx-rebirth/issues/724>