Update generate-kconfig-udlr.py for Python AST deprecation
Python 3.8 deprecated use of `Num`, replacing it with `Constant`.
Python 3.12 issues a runtime warning for use of the old name, and
indicates that Python 3.14 will drop support for the old form.  There is
no need to use the old form, so convert to the new form.  Add a type
check that rejects use of non-numbers here, since those will fail later.
Rejecting it here produces a clear error message.