Add command line option to adjust render zoom level
Add experimental new option `-render-zoom INTEGER`.  INTEGER must be a
decimal integer in the range of [`-36824`, `INT_MAX`].  In competitive
multiplayer games, this option is ignored.  In game modes other than
competitive multiplayer, INTEGER is added to the normal render zoom
factor of `36864` (`0x9000`).  The renderer tends toward excessive
culling when the zoom level gets very small, so any sum less than 40 is
treated as 40.  No upper bound is set, but very large values will likely
make the game unplayable due to visual distortion.

If the command line option is not used, then INTEGER is treated as 0,
and the game renders as it does in builds which do not define this
option.

When the option is used, negative values reduce the viewable area and
give the appearance that the viewer closer to the target.  Positive
values increase the viewable area and give the appearance of being
further from the target.

The effect of this option is cumulative with the transient changes that
developer builds can apply to the `Render_zoom` variable.

This implementation was aided by the hints from JeodC and btb about how
the renderer uses Render_zoom to influence the field-of-view.

Requested-by: Q3BFG10K <https://github.com/dxx-rebirth/dxx-rebirth/issues/775>
Thanks-to: JeodC <https://github.com/dxx-rebirth/dxx-rebirth/issues/775#issuecomment-2094362796>
Thanks-to: btb <https://github.com/dxx-rebirth/dxx-rebirth/issues/775#issuecomment-2094364294>