Add explicit length limit on movie filename
Some early releases of gcc-13.2 compute an incorrectly pessimistic
outcome for `BuiltinMovies::BuiltinMovies` constructing movie filenames,
and assume that the entire size of `movielib_files` can be written into
the buffer, even though `movielib_files` is a
`constexpr std::array<std::array<char, 8>, 3>` and each of the inner
arrays is individually null terminated.

Add an explicit length limit (`%.8s`) on the string to remind gcc that a
string consisting of 7 bytes plus a null terminator cannot produce more
than 8 characters.  While in the area, add a check on the return value
of `snprintf` and refuse to use a buffer that was truncated, even though
such truncation should be impossible now.

Reported-by: rtissera <https://github.com/dxx-rebirth/dxx-rebirth/issues/778>