Include presence of sharepath in default build directory checksum
The value of sharepath is only defined for a few files, but most files
need to be told whether sharepath is empty or non-empty, since that
affects the definition of CArg.  Mix that into the string used to derive
the build directory path.  Without that, the following configuration
fails:

- scons d1x_sharepath= d2x_sharepath=/non-empty/ lto=1 builddir_prefix=build/

The mismatch of sharepath-disabled for D1 and sharepath-enabled for D2
should cause the common files to be built in different directories.
Prior to this commit, it does not, so then D2 links in common files
built with sharepath-disabled and non-common files with
sharepath-enabled.  This is always wrong, and when LTO is used, the
linker flags it as an error.

With this commit, the two games compute different paths for their files,
and the common files are built separately for D1 versus D2, allowing D2
to get sharepath-enabled common files.