Move -Wformat=2 out of the always-Werror block
Some warnings are always promoted to error via `get_Werror_sequence`.
Others are left as a warning, and rely on the user to add `-Werror` to
promote them to an error.  `-Wformat=2` has been in the always-error
group, but periodically causes problems due to new compiler versions
changing the specific circumstances under which they issue format string
warnings.  Move -Wformat=2 to the only-warnings block to avoid user
builds breaking unexpectedly.  My personal configuration still sets
`-Werror`, and so will continue to treat this as an error.