Only set -Werror= for specific configure tests
Previously, selected well-supported -Wfoo flags were raised to
-Werror=foo if the user did not enable -Werror in the CXXFLAGS for a
build.  When the user set -Werror, then -Wfoo was kept as-is.  This was
intended to ensure a consistent level of warnings across all developers,
and discourage submitting code that generated common warnings.
Contributions have been rare, and generally small enough that even if a
warning were introduced, fixing it as part of the merge would be viable.
Per request from a Gentoo maintainer, remove the automatic -Werror=foo
conversion.  Rely on interested users adding -Werror to the CXXFLAGS for
their build.

Requested-by: chewi <https://github.com/dxx-rebirth/dxx-rebirth/issues/790#issue-2373773977>