Tighten check for sufficient songs
The previous check would reject playing a song if exactly zero songs
were available.  This is insufficient.  A `.sng` file that defines
exactly 1 song will pass the preceding check, but then index out of
bounds when trying to play `song_number::briefing`.  Change the check to
require that the index is in range.

Reported-by: BigheadSMZ <https://github.com/dxx-rebirth/dxx-rebirth/issues/768#issuecomment-2188084417>