Simplify PHYSFSX_fgets test for newline
Switch from incrementing `p`, then conditionally decrementing `p` on not
finding a newline to instead create an incremented temporary, updating
`p` on finding a newline, and leaving `p` unchanged on the lack of a
newline.  This seems easier to read, and has the minor improvement that
the no-newline path is now empty.