Raise heuristic out of _result_check_user_setting
Switch from passing a value that is either a string (to be defined equal
to the truth of the condition) or a tuple (to be defined if and only if
the condition is true) to instead pass two variables, one for each
purpose.  Update callers to use the argument appropriate to what the
caller wants.