xrange: add requires() clauses to some methods
Add requires() to operator-(xrange_iterator) so that callers can detect
when operator- is unsupported for a given `index_type`.

Add requires() to operator++() and split it into three methods,
depending on whether the step is increment, decrement, or addition.

Add requires() to operator++(int) so that callers can detect whether
operator++() is supported.