Fix D1 weapon_info parsing
The serialization path for `d1x::weapon_info` lacked a `static_assert`
to check that the size was correct.  This omission allowed a size change
in commit 49bc677729333064317e4a91fd57e960525ad1f2 to go undetected,
resulting in the game reading the wrong number of bytes from the file
each time it tried to deserialize a `d1x::weapon_info`.  In turn, that
caused the loaded `weapon_info` to be invalid, and caused secondary
errors elsewhere as the incorrect file position lead to incorrect loads
of other data structures.

Add the missing `serial::pad<1>()` uses that should have been in
commit 49bc677729333064317e4a91fd57e960525ad1f2.  Also, add a static
assertion on the size, so that any future mistakes are caught at compile
time.

Reported-by: adolfogp <https://github.com/dxx-rebirth/dxx-rebirth/issues/807>
Fixes: 49bc677729333064317e4a91fd57e960525ad1f2 ("Use `sound_effect` for arguments and members, where appropriate")