Change `savegame_pof_names_type` length to MAX_POLYGON_MODELS for D1
As released, Descent 1 used `MAX_POLYGON_MODELS_NEW` (defined to 167) as
the length of the Save_pof_names array.  However, the code to load that
array contained `Assert(N_save_pof_names < MAX_POLYGON_MODELS);`, so in
practice, any file containing more than `MAX_POLYGON_MODELS` would fail
the assertion in debug-enabled builds.

Descent 2 already used `MAX_POLYGON_MODELS` as the length of
`Save_pof_names`.

Commit 8b713978fcdbb0433ddcb57387f19fee565f9fdd ("Ignore bogus POF
counts") replaced the `Assert` with a runtime check and associated
diagnostic, so any invalid levels have failed to load their POF data
(even if the load would not cause a crash) since 2017, and no one has
reported this as a problem.

Reduce the length to `MAX_POLYGON_MODELS` in Descent 1, to match the
length used in Descent 2 and to match the length constraint imposed
first by the `Assert` and later by the runtime check.