Remove obsolete workaround for gcc bug #71740
Per upstream notes, this bug is fixed in gcc-11.  The minimum supported
gcc for Rebirth is gcc-12, so all supported versions of gcc include the
fix.  Additionally, the workaround was guarded by a preprocessor check
to apply only to gcc-6 through gcc-10, and no higher.  Even if the bug
were not fixed upstream, the preprocessor guard would prevent the
workaround from applying to any supported version of gcc, and so the
workaround is now useless as written.  Remove it.