Simplify `gamefont_loadfont`
The test for file existence is unnecessary.  There are two callers of
`gamefont_loadfont`.
- `gamefont_choose_game_font` can only use entries that were previously
  initialized, so if the font existed when it was initialized, assume it
  still exists now.
- `addfontconf` already tests whether the font file exists, before
  calling `gamefont_loadfont`, so there is no need to retest the
  existence in `gamefont_loadfont`.
Based on that, remove the test for existence, and the special logic
that would run if the file did not exist.

On the existence path, remove the call to `gamefont_unloadfont`.  This
call would reset `.cur`, which gets rewritten again before the end of
`gamefont_loadfont`, and clears the font pointer, which gets reset with
the value of `gr_init_font`.  Therefore, the only value to the call to
`gamefont_unloadfont` is that it would free the old font before
allocating the new one, rather than after.