Simplify loading D1 fonts
For each slot in `fc.fontconf`, there are 3 callers of `addfontconf` in
Descent 1, and 2 in Descent 2.  The first caller is always for 640x480.
The last caller is always for 320x200, and cannot collide with anything.
In Descent 1, there is a middle caller which can collide with the first
caller, if conditions are met.  When this collision occurs, the work
from the first call is reverted, and the font from the middle call is
used.  In Descent 2, there is no middle caller, so with one call for
640x480 and one for 320x200, there is never a collision, and the test
for `f.expected_screen_resolution_x == expected_screen_resolution_x` is
always false.

Previously, Descent 1 loaded the non-preferred (and usually absent[1])
high-resolution font first, then if Mac data was available, tried to
replace the non-preferred font with a preferred high-resolution font
from Mac data.  Remove the replacement logic, and instead arrange for
Descent 1 to load the preferred font first, with a fallback to the
non-preferred font.  The Descent 2 logic should be the same before this
commit as after.

[1]: The high-resolution font is not present in Descent 1 shareware or
Descent 1 retail data.  However, for the benefit of users using windows
640x480 or larger, D1X-Rebirth tries to load Descent 2 high-resolution
fonts if the user has made them available by adding them to the game's
search path.  When absent, as they usually are, D1X-Rebirth falls back
to using D1 fonts.