Recover from failure to find adequate size font
In theory, safety code elsewhere should normally prevent the user from
picking a screen size that triggers this error.  However, if the user
has a custom set of data files that omits support for 320x200, and the
user resizes to a window smaller than 640x480, then this path could
trigger.  Instead of terminating the entire program, use default scaling
for fonts, and log a message to the user.  Since the user can trigger a
resolution change, and corresponding font rescan, at any time, the user
might be in the middle of an unsaved game when this happens.
Terminating the entire program with unsaved state is needlessly
disruptive.