songs_play_level_song: cast directly to song_number
`size` is typically `std::size_t`.  On 64-bit platforms, `std::size_t`
is typically a 64-bit value.  As a result, `(-levelnum - 1) % size` is a
64-bit result, even though `levelnum` is `int` (and thus usually 32-bits
wide).  For this reason, a `static_cast<unsigned>` was used to discard
the unused upper zero bits, so that `song_number{expr}` would succeed.

On 32-bit systems, such as Win32 and some ARM systems, `std::size_t` is
a 32-bit value, so `(-levelnum - 1) % size` is 32-bits, and the
`static_cast<unsigned>` is useless, and provokes a `-Wuseless-cast`
warning.

Switch to directly using `static_cast<song_number>`.  It is exactly as
safe as casting to unsigned (which is narrowing on 64-bit platforms, but
is safe due to the limited legal input values) and then doing a
non-narrowing conversion to `song_number`.

Reported-by: joolswills <https://github.com/dxx-rebirth/dxx-rebirth/issues/813> [ARM / Raspberry Pi 5]