Pass simpler types to g3 drawing operations
Introduce type aliases `g3_draw_line_point`, `g3_draw_sphere_point`, for
the point used for drawing a line or sphere, respectively.  Adjust the
prototypes for the called functions to use these aliases.  In the OpenGL
case, set the alias to a base of `g3s_point`, since the called function
does not need the full data of a `g3s_point`.