README.md: remove reference to zico's e-mail address
In early 2024, zico changed the Rebirth website not to offer snapshots
of the current git code, and closed his GitHub account.  The site now
reads, in part:

```
About 10 years later, when I became increasingly involved in other projects, I just couldn't dedicate the time to this port that it and its community deserved anymore and in 2020, I decided to officially retire.
```

Therefore, directing people to contact him about issues with the current
code is not helpful, as he has no access to correct any problems
reported, and likely does not keep current on the state of the project.
His work over many years is appreciated, and he should be left to enjoy
his retirement.  Direct any future readers to the GitHub repository.