Only send boss gate messages if in multiplayer
This is the same bug as with the automap msgsend state fixed earlier.
The outbound queue slowly fills with boss gate messages, eventually
leading to a crash.  The bug is easily detected by the new assert in
`dispatch_table::send_data`, which reports an assertion failure on the
first boss gate message in single player.