Only set automap msgsend state if in a multiplayer game
Setting the msgsend state queues a message to be sent to other players.
If the game is in single player mode, the queue is never drained, so
repeated cycling of the automap slowly fills the queue, eventually
leading to a crash.  Change the logic not to queue a msgsend state
message if not in multiplayer.

Reported-by: seanwlx <https://github.com/dxx-rebirth/dxx-rebirth/issues/817>