Only send guidebot message if in multiplayer
This is the same bug as with the automap msgsend state fixed earlier.
The outbound queue slowly fills as the player issues guidebot orders,
eventually leading to a crash.  The bug is easily detected by the new
assert in `dispatch_table::send_data`, which reports an assertion
failure on the first explicit command to the guidebot in single player.