Only send decloak message if in multiplayer
This is the same bug as with the automap msgsend state fixed earlier,
although it is much harder to trigger the crash since the player would
need to repeatedly cloak and decloak to fill the outbound queue.  The
bug is easily detected by the new assert in `dispatch_table::send_data`,
which reports an assertion failure on the first decloak in single
player.