`trace_segs`: simplify picking next side to check
Observe that `build_side_dists` only sets a bit in `.centermask` if a
non-zero value is written to `.distances`, and that `.distances` is
value-initialized.  Therefore, if the bit in `centermask` is clear, the
corresponding element in `.distances` retains its original zero from the
value-initialization.  Since the test in `trace_segs` seeks sides for
which the value in `.distances` is less than `biggest_val`, and
`biggest_val` is initialized to `0`, any side N for which
`.centermask` is unset will have its `.distances[N]` retain its
value-initialized value.  Such a value will be rejected by the less-than
test.  Remove the redundant test for `centermask`.