`draw_all_edges`: use `std::ranges::sort`
Switch to `std::ranges::sort` and use a projection on the element to
convert from the `const Edge_info *` supplied by the to-be-sorted range
to the rotated z coordinate of that vertex.  This allows more of the
comparison logic to remain in the standard library, and produces
slightly smaller code.