Avoid exception on invalid number of fuel centers
Previously, the use of `at()` would trigger the standard library to
raise an exception if the fuel center index was out of range.  Switch to
validate the index first, and silently return if the index is invalid.
Invalid indexes should never happen, so the change should not be
noticeable to users.