Add `robot_gun_goal_index` for ai_local::goal_state, achieved_state
`ai_local::goal_state` and `ai_local::achieved_state` are
`MAX_SUBMODELS` elements long.  All other arrays indexed by
`robot_gun_number` are `MAX_GUNS` long.  This inconsistency interferes
with assuming that a given value for an enum used to index
enumerated_array is either always valid or never valid, since
an enum could be valid for a MAX_SUBMODELS array (10 elements), but
invalid for a MAX_GUNS array (8 elements).

Add a dedicated index type for use with these arrays, and add
appropriate casts from the regular index type.  The unusual length may
be a bug, but does not appear to hurt the program, and proving it to be
a bug requires more detailed analysis.  For now, apply the split type so
that all arrays can keep their traditional lengths.

These changes have no effect on the generated code.  The preceding
commit and this one generate the same program.