Show version of git during SConf setup
SCons does not use the version for any purpose, but it may be useful
when reviewing user-submitted build logs to see what version of git (if
any) was available.  Users building from an archive, or in a restricted
container, might not have git.  Absence of git is not a fatal, but does
impair certain other steps that have been present for a long time.