Pass generated kconfig UDLR header via `-include`
Switch from a `#include` in `kconfig.cpp` dependent on
`DXX_BUILD_DESCENT` to pass the header via a `-include` on the compiler
command line.  This ensures that the file includes the header that SCons
just generated, rather than potentially finding a header of that name
elsewhere on the include search path.

Rename `StaticObject_hook` to `StaticObject_wrapper`, and change the
semantic to be that it is called with the inputs for a call to
`StaticObject`, which it then uses to make the call, so that it has
access to both the inputs and the output.