SConstruct: only report unknown variables from command line
scons-4.9 fixed <https://github.com/SCons/scons/issues/4645>, causing
`Variables.UnknownVariables()` to report unknown variables defined in
files.  Rebirth SConstruct relied on the older behavior, so that the
user could define variables in a site file that would be registered and
used only if requested by the command line enabling the associated
profile.  Prior to scons-4.9, the below worked:

```python
	# site-local.py
gcc14_CXXFLAGS = '-Werror'
```
```
$ scons dxx=gcc14,	# OK, finds and uses -Werror
$ scons dxx=gcc13,	# OK, ignores the gcc-14 specific -Werror
```

Starting in scons-4.9, the `dxx=gcc13,` run fails because Rebirth
SConstruct registers gcc13_CXXFLAGS (which is not set by the file) but
does not register gcc14_CXXFLAGS (because no profile named gcc14 is
requested).  The check for unknown variables then traps because
>=scons-4.9 returns the gcc14_CXXFLAGS from `site-local.py` as unknown.

Avoid the diagnostic by filtering the set of unknown variables down to
only those that are unknown and are in ARGUMENTS.  Any unknown variables
that are only set in `site-local.py` are then not reported.