`paging_touch_side`: remove call to `texmerge_get_cached_bitmap`
`texmerge_get_cached_bitmap` composes the two input textures into a
bitmap, puts the bitmap in the cache for future use, and returns the
bitmap.  `paging_touch_side` discarded the bitmap, so the only value to
calling `texmerge_get_cached_bitmap` was to populate the cache.  The
cache is small, so some levels might fill the cache and begin discarding
elements before all the calls to `paging_touch_side` were done.  Remove
the call to `texmerge_get_cached_bitmap` and let the code that needs
merged bitmaps create them as needed.